+++ src/test/java/com/radello/constructioncompanyorganizer/converter/IndicativeCostCommandToIndicativeCostTest.java
@@ -11,7 +11,7 @@ import static org.junit.jupiter.api.Assertions.*;
 class IndicativeCostCommandToIndicativeCostTest {
 
     private final String FOR_WHAT_VALUE = "Flat";
-    private final int ID_VALUE = 1;
+    private final Long ID_VALUE = 1L;
 
     IndicativeCostCommandToIndicativeCost indicativeCostCommandToIndicativeCost;
