+++ src/test/java/com/radello/constructioncompanyorganizer/domain/IndicativeCostTest.java
@@ -17,7 +17,7 @@ class IndicativeCostTest {
     @Test
     void getID() {
 
-        int testId = 1;
+        Long testId = 1L;
 
         indicativeCost.setID(testId);
